# docs/swagger.yaml
basePath: /api
definitions:
  courseerror.CourseError:
    properties:
      code:
        type: integer
      error:
        type: string
    type: object
...
  entity.Credentials:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
...
  entity.PasswordRecoverCredentials:
    properties:
      code:
        type: integer
      email:
        type: string
      password:
        type: string
    type: object
...
  entity.SuccessResponse:
    properties:
      message:
        type: string
      success:
        type: boolean
    type: object
...
host: localhost:70
info:
  contact: {}
  title: Приложение course
  version: "1.0"
paths:
...
  /v1/auth/email/newConfirmKey:
    get:
      parameters:
      - description: почта для отправки кода подтверждения
        in: query
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.SuccessResponse'
        "400":
          description: Провалена валидация или почта пользователя уже подтверждена
          schema:
            $ref: '#/definitions/courseerror.CourseError'
        "500":
          description: Возникла внутренняя ошибка
          schema:
            $ref: '#/definitions/courseerror.CourseError'
      summary: Отправить новый код подтверждения
      tags:
      - Методы для авторизации пользователей
  /v1/auth/email/verification:
    post:
      parameters:
      - description: Код подтверждения
        in: query
        name: confirmCode
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.SuccessResponse'
        "400":
          description: Провалена валидация, или декодирование сообщения, или почта
            пользователя уже подтверждена
          schema:
            $ref: '#/definitions/courseerror.CourseError'
        "403":
          description: Код подтверждения не совпал
          schema:
            $ref: '#/definitions/courseerror.CourseError'
        "404":
          description: Код подтверждения не найден
          schema:
            $ref: '#/definitions/courseerror.CourseError'
        "500":
          description: Возникла внутренняя ошибка
          schema:
            $ref: '#/definitions/courseerror.CourseError'
      summary: Подтвердить почту пользователя
      tags:
      - Методы для авторизации пользователей
  /v1/auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: учетные данные
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/entity.Credentials'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.SuccessResponse'
        "400":
          description: Провалена валидация, или декодирование сообщения, или почта
            уже занята
          schema:
            $ref: '#/definitions/courseerror.CourseError'
        "404":
          description: Пользователь не найден
          schema:
            $ref: '#/definitions/courseerror.CourseError'
        "405":
          description: Пользователь неактивен
          schema:
            $ref: '#/definitions/courseerror.CourseError'
        "500":
          description: Возникла внутренняя ошибка
          schema:
            $ref: '#/definitions/courseerror.CourseError'
      summary: Залогиниться пользователю
      tags:
      - Методы для авторизации пользователей
  /v1/auth/recoverPassword:
    post:
      consumes:
      - application/json
      parameters:
      - description: почта, новый пароль и код
        in: body
        name: recoverCredentials
        required: true
        schema:
          $ref: '#/definitions/entity.PasswordRecoverCredentials'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.SuccessResponse'
        "400":
          description: Провалена валидация или код подтверждения неверный
          schema:
            $ref: '#/definitions/courseerror.CourseError'
        "404":
          description: Код не найден
          schema:
            $ref: '#/definitions/courseerror.CourseError'
        "500":
          description: Возникла внутренняя ошибка
          schema:
            $ref: '#/definitions/courseerror.CourseError'
      summary: Установить новый пароль
      tags:
      - Методы для авторизации пользователей
  /v1/auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: учетные данные
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/entity.Credentials'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.SuccessResponse'
        "400":
          description: Провалена валидация, или не удалось декодировать сообщение,
            или почта уже занята
          schema:
            $ref: '#/definitions/courseerror.CourseError'
        "500":
          description: Возникла внутренняя ошибка
          schema:
            $ref: '#/definitions/courseerror.CourseError'
      summary: Зарегестрироваться пользователю
      tags:
      - Методы для авторизации пользователей
  /v1/auth/sendRecoveryCode:
    get:
      parameters:
      - description: почта для восстановления пароля
        in: query
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.SuccessResponse'
        "400":
          description: Провалена валидация
          schema:
            $ref: '#/definitions/courseerror.CourseError'
        "500":
          description: Возникла внутренняя ошибка
          schema:
            $ref: '#/definitions/courseerror.CourseError'
      summary: Отправить код для восстановления пароля
      tags:
      - Методы для авторизации пользователей
...
securityDefinitions:
  AdminApiKeyAuth:
    in: cookie
    name: admin_auth
    type: apiKey
  ApiKeyAuth:
    in: cookie
    name: auth
    type: apiKey
swagger: "2.0"